Cadastro de vaga ok

+++ ventura_rh/lib/models/vaga/vaga.dart
@@ -28,25 +28,23 @@ class Vaga extends ChangeNotifier {
 
   Vaga(
       {this.id,
-        this.userId,
-        this.user,
-        this.companyTitle,
-        this.titleVacancy,
-        this.descriptionVacancy,
-        this.images,
-        this.workplace,
-        this.workload,
-        this.numberOfVacancies,
-        this.address,
-        this.criterios,
-        this.newImages,
-        this.deleted = false}) {
+      this.userId,
+      this.user,
+      this.companyTitle,
+      this.titleVacancy,
+      this.descriptionVacancy,
+      this.images,
+      this.workplace,
+      this.workload,
+      this.numberOfVacancies,
+      this.address,
+      this.criterios,
+      this.newImages,
+      this.deleted = false}) {
     images = images ?? [];
-    criterios =  criterios ?? [];
+    criterios = criterios ?? [];
     newImages = newImages ?? [];
   }
-
-
 
   final Firestore firestore = Firestore.instance;
   final FirebaseStorage storage = FirebaseStorage.instance;
@@ -67,19 +65,20 @@ class Vaga extends ChangeNotifier {
   int numberOfVacancies;
   Address address;
   List<Criterio> criterios;
-  List<String> newImages;
+  List<dynamic> newImages;
   File file;
   bool deleted;
 
   bool _loading = false;
+
   bool get loading => _loading;
-  set loading(bool value){
+
+  set loading(bool value) {
     _loading = value;
     notifyListeners();
   }
 
-
-  List<Map<String, dynamic>> exportSizeList(){
+  List<Map<String, dynamic>> exportCriterioList() {
     return criterios.map((c) => c.toMap()).toList();
   }
 
@@ -91,48 +90,48 @@ class Vaga extends ChangeNotifier {
       'userId': userId,
       'address': address.toMap(),
       'companyTitle': companyTitle,
-      'titleVacancy':titleVacancy,
-      'descriptionVacancy':descriptionVacancy,
-      'image':images,
-      'workplace':workplace,
-      'workload':workload,
-      'numberOfVacancies':numberOfVacancies,
-      'criterio':exportSizeList()
+      'titleVacancy': titleVacancy,
+      'descriptionVacancy': descriptionVacancy,
+      'images': images,
+      'workplace': workplace,
+      'workload': workload,
+      'numberOfVacancies': numberOfVacancies,
+      'criteria': exportCriterioList()
     };
 
-     List<String> updateImages = [] ;
-
-
-    for(final newImage in newImages)
-    { if(!images.contains(newImage)){
-
-        final StorageUploadTask task = storageRef.child(Uuid().v1()).putFile(newImage as File);
-        final StorageTaskSnapshot snapshot = await task.onComplete;
-        final String url = await snapshot.ref.getDownloadURL() as String;
-        updateImages.add(url);
-      }}
-
-
-
-    if(id == null){
+    if (id == null) {
       final doc = await firestore.collection('vaga').add(data);
       id = doc.documentID;
+      firestoreRef.updateData({'id':id});
     } else {
       await firestoreRef.updateData(data);
     }
 
+    List<String> updateImages = [];
 
-      //removendo do firebase imagem
-         for(final image in images)
-         {if(!newImages.contains(image) && images.contains('firebase')){
-      try {
-        final ref = await storage.getReferenceFromUrl(image);
-        await ref.delete();
-      } catch (e){
-        debugPrint('Falha ao deletar $image');
+    for (final newImage in newImages) {
+      if (images.contains(newImage)) {
+        updateImages.add(newImage as String);
+      } else {
+        final StorageUploadTask task =
+            storageRef.child(Uuid().v1()).putFile(newImage as File);
+        final StorageTaskSnapshot snapshot = await task.onComplete;
+        final String url = await snapshot.ref.getDownloadURL() as String;
+        updateImages.add(url);
       }
-    }}
+    }
 
+    //removendo do firebase imagem
+    for (final image in images) {
+      if (!newImages.contains(image) && images.contains('firebase')) {
+        try {
+          final ref = await storage.getReferenceFromUrl(image);
+          await ref.delete();
+        } catch (e) {
+          debugPrint('Falha ao deletar $image');
+        }
+      }
+    }
 
     await firestoreRef.updateData({'images': updateImages});
 
@@ -141,26 +140,28 @@ class Vaga extends ChangeNotifier {
     loading = false;
   }
 
-
-  void delete(){
+  void delete() {
     firestoreRef.updateData({'deleted': true});
   }
 
-
-  Vaga clone(){
+  Vaga clone() {
     return Vaga(
       id: id,
       userId: userId,
-      companyTitle:companyTitle,
-      titleVacancy:titleVacancy,
-      descriptionVacancy:descriptionVacancy,
-      workplace:workplace,
-      workload:workload,
-      numberOfVacancies:numberOfVacancies,
+      companyTitle: companyTitle,
+      titleVacancy: titleVacancy,
+      descriptionVacancy: descriptionVacancy,
+      workplace: workplace,
+      workload: workload,
+      numberOfVacancies: numberOfVacancies,
       images: List.from(images),
       criterios: criterios.map((c) => c.clone()).toList(),
       address: address,
-
     );
+  }
+
+  @override
+  String toString() {
+    return 'Vaga{firestore: $firestore, storage: $storage, id: $id, userId: $userId, user: $user, companyTitle: $companyTitle, titleVacancy: $titleVacancy, descriptionVacancy: $descriptionVacancy, images: $images, workplace: $workplace, workload: $workload, numberOfVacancies: $numberOfVacancies, address: $address, criterios: $criterios, newImages: $newImages, file: $file, deleted: $deleted, _loading: $_loading}';
   }
 }
